LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/selection/contenteditable/modifying-selection-with-non-primary-mouse-button.tentative_secondary-expected.txt

CONSOLE MESSAGE: Error: assert_equals: selection should be collapsed in the second <span> element which was clicked by secondary button before pointerup event expected Text node "second span." but got Text node "first span."
CONSOLE MESSAGE: Error: assert_equals: Selection should be collapsed in the second <span> element which was clicked by secondary button before mouseup event expected Text node "second span." but got Text node "first span."
first span.
second span.
link.

Harness Error (FAIL), message = Error: assert_equals: Selection should be collapsed in the second <span> element which was clicked by secondary button before mouseup event expected Text node "second span." but got Text node "first span."

PASS secondary click should set focus to clicked editable element and collapse selection around the clicked point
PASS secondary click should move caret in an editable element
PASS secondary click shouldn't move caret in an editable element if the default of mousedown event is prevented
PASS secondary click shouldn't move caret in an editable element if the default of pointerdown event is prevented
FAIL Shift + secondary click should extend the selection promise_test: Unhandled rejection with value: object "Error: testdriver-vendor.js for WebKit does not yet support mixing key and pointer sources"
FAIL Shift + secondary click in a link shouldn't extend the selection promise_test: Unhandled rejection with value: object "Error: testdriver-vendor.js for WebKit does not yet support mixing key and pointer sources"
FAIL secondary mouse button down should move caret, but its button up shouldn't move caret assert_equals: Selection#focusNode should stay in the second <span> element expected Text node "second span." but got Text node "first span."
